+++ paso/learners/learners.py
@@ -4,7 +4,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib import rcParams
 
-#from tqdm import tqdm
+# from tqdm import tqdm
 
 import warnings
 
@@ -14,21 +14,22 @@ from sklearn.metrics import f1_score, accuracy_score
 from sklearn.metrics import precision_score, log_loss, recall_score
 from sklearn.metrics import confusion_matrix
 from sklearn.metrics import roc_auc_score
-#from sklearn.metrics import brier_score_loss
+
+# from sklearn.metrics import brier_score_loss
 
 # paso imports
-from paso.base import pasoModel, NameToClass
-from paso.base import raise_PasoError
-from paso.base import _array_to_string, pasoDecorators
-from paso.base import _add_dicts
-#from paso.base import _dict_value2
+from paso.base import pasoModel, NameToClass, _check_non_optional_kw
+from paso.base import raise_PasoError, _exists_as_dict_value, _dict_value
+from paso.base import _array_to_string, pasoDecorators, _kind_name_keys
+from paso.base import _add_dicts, _exists_Attribute
+
+
+# from paso.base import _dict_value2
 from loguru import logger
 
 __author__ = "Bruce_H_Cottman"
 __license__ = "MIT License"
-#
-
-
+# class Learner
 class Learner(pasoModel):
     """
     Currently, **paso** supports only the sklearn RandomForest
@@ -54,7 +55,6 @@ class Learner(pasoModel):
 
         """
         super().__init__()
-        self.inplace = False
 
     def learners(self):
         """
@@ -66,8 +66,8 @@ class Learner(pasoModel):
         """
         return list(NameToClass.__learners__.keys())
 
-    @pasoDecorators.TrainWrapsklearn(array=False)
-    def train(self, X, **kwargs):
+    @pasoDecorators.TTWrap(array=False)
+    def train(self, X, *args, **kwargs):
         # Todo:Rapids numpy
         """
         Parameters:
@@ -75,65 +75,71 @@ class Learner(pasoModel):
         Returns:
             self
         """
+        # currently support only one learner, very brittle parser
+        if self.kind == {}:
+            raise_PasoError(
+                "keyword kind must be present at top level:{}:".format(
+                    self.ontology_kwargs
+                )
+            )
+        if _exists_Attribute(self, "target"):
+            if self.target in self.Xcolumns:
+                self.y_train = X[self.target].values
+                self.X_train = X[X.columns.difference([self.target])].values
+            # no target set,
+            else:
+                raise_PasoError(
+                    "learner:fit: unknown target:{} in {} for learner: {}".format(
+                        self.target, self.Xcolumns, self.kind_name
+                    )
+                )
+        else:
+            raise_PasoError(
+                "learner:train: target: in must be a keyword in call to train: {}".format(
+                    kwargs
+                )
+            )
 
-        # todo refactor using non-optional-kw
-        # checks for non-optional keyword
-        if self.learner not in NameToClass.__learners__:
-            raise_PasoError(
-                "TrainWrap: No learner named: {} not in learners;: {}".format(
-                    self.learner, NameToClass.__learners_.keys()
-                )
-            )
-        if self.cross_validation == None:
-            raise_PasoError(
-                "cross_validation not specified for learner: {}".format(self.learner)
-            )
-        if self.target == None:
-            logger.warning(
-                "target not specified through keyword call or ontological file for learner: {}".format(
-                    self.target
-                )
-            )
-        if self.target not in self.Xcolumns:
-            raise_PasoError(
-                "trainwrap: unknown target:{} in {}".format(self.target, self.Xcolumns)
-            )
+        # list(self.kind.keys())[0]
+        # self.kind_name_key = _exists_as_dict_value(self.kind, self.kind_name)
+        # # Thus order keyword must be given even if []
+        # kind_name_kwargs = _kind_name_keys(
+        #     self, self.kind_name_key, verbose=self.verbose
+        # )
+
         # create instance of this particular learner
-        if self.learner_name_kwargs == None:
-            self.model = NameToClass.__learners__[self.learner]()
-            logger.warning(
-                "model_kwargs not specified for learner: {}".format(self.modelKey)
+        # checks for non-optional keyword
+        if self.kind_name not in NameToClass.__learners__:
+            raise_PasoError(
+                "Train; no operation named: {} not in learners;: {}".format(
+                    self.kind_name, NameToClass.__learners__.keys()
+                )
             )
         else:
-            self.model = NameToClass.__learners__[self.learner](**self.kwargs)
+            self.model = NameToClass.__learners__[self.kind_name](
+                **self.kind_name_kwargs
+            )
+            self.model.fit(self.X_train, self.y_train)
+            self.clf_type = self.type
+
+        # operation on this class by keyword order
+        kwa = "order"
+        order_ops = _dict_value(
+            self.kind_name_keys, kwa, []
+        )  # should be a lis or error, no constant
+        for id_ops, ops_name in enumerate(order_ops):
+            ops_name_key = _exists_as_dict_value(self.kind_name_keys, ops_name)
+            ops_name_kwargs = _kind_name_keys(self, ops_name_key, verbose=self.verbose)
             if self.verbose:
                 logger.info(
-                    "learner: {} with kwargs {}".format(self.learner, self.kwargs)
+                    "\n clf-ops-name:{}\n kwargs: {}".format(ops_name, ops_name_kwargs)
                 )
-
-        # X_train,y_train turned into arrays from dataframe, used in predict
-        if self.target in self.Xcolumns:
-            self.n_class = X[self.target].nunique()
-            self.class_names = _array_to_string(X[self.target].unique())
-
-        # todo support other cross-valid
-        # todo  support named metrics
-        self.y_train = X[self.target].values
-        self.X_train = X[X.columns.difference([self.target])]
-        self.clf = self.model  # start with base learner
-        for i, cv_name in enumerate(self.cross_validation):
-            if self.verbose:
-                logger.info("cross_validation: {}".format(cv_name))
-
-            cv_name_kwargs = self.cv_kwargs_list[i]
-            if self.verbose:
-                logger.info("    cv kwargs: {}".format(cv_name_kwargs))
-            self.clf = NameToClass.__cross_validators__[cv_name](
-                self.clf, **cv_name_kwargs
+            self.model = NameToClass.__learners__[ops_name](
+                self.model, **ops_name_kwargs
             )
-            self.clf.fit(self.X_train, self.y_train)
+            self.model.fit(self.X_train, self.y_train)
 
-        return self.clf
+        return self
 
     @pasoDecorators.PredictWrap(array=False, narg=2)
     def predict(self, X, **kwargs):
@@ -147,57 +153,57 @@ class Learner(pasoModel):
         Returns:
             (DataFrame): predict of X
         """
-        if self.measure:
-            self.y_test = X[self.target].values
-            self.X_test = X[X.columns.difference([self.target])]
-
-        y_pred = self.clf.predict(self.X_test)
-
-        if self.measure:
-            self.wrong_predicted_class = X[X[self.target] != y_pred]  # df
-
-        y_predp = self.clf.predict_proba(self.X_test)
-
-        if self.type == "Classification":
+        # assured by predictwrap that train has been called already
+        _check_non_optional_kw(
+            self.measure, "measure kw must be present in call to predict"
+        )
+        if self.clf_type == "Classification":
             if self.measure:
-                self.metrics = _add_dicts(
-                    self.__metric_class(y_pred, self.y_test),
-                    self.__metric_class_probability(y_predp, self.y_test),
-                )
-
-                if self.plot_confusion_matrix and self.verbose:
-                    cm = confusion_matrix(self.y_test, y_pred)
-                    self.metrics = _add_dicts(self.metrics, {"confusion_matrix": cm})
+                self.n_class = X[self.target].nunique()
+                self.class_names = _array_to_string(X[self.target].unique())
+                self.y_test = X[self.target].values
+                self.X_test = X[X.columns.difference([self.target])].values
+                y_pred = self.model.predict(self.X_test)
+                self.wrong_predicted_class = X[X[self.target] != y_pred]  # df
+                y_predp = self.model.predict_proba(self.X_test)
+                logger.debug("type:{}.".format(self.type))
+                if self.clf_type == "Classification":
+                    self.metrics = _add_dicts(
+                        self.__metric_class(y_pred, self.y_test),
+                        self.__metric_class_probability(y_predp, self.y_test),
+                    )
                     if self.verbose:
-                        # Plot non-normalized confusion matrix
-                        self._plot_confusion_matrix(
-                            cm,
-                            self.class_names,
-                            normalize=False,
-                            title="Confusion matrix, without normalization",
+                        cm = confusion_matrix(self.y_test, y_pred)
+                        self.metrics = _add_dicts(
+                            self.metrics, {"confusion_matrix": cm}
                         )
-                        # Plot normalized confusion matrix
-                        self._plot_confusion_matrix(
-                            cm,
-                            self.class_names,
-                            normalize=True,
-                            title="Normalized confusion matrix",
+                        if self.verbose:
+                            # Plot non-normalized confusion matrix
+                            self._plot_confusion_matrix(
+                                cm,
+                                self.class_names,
+                                normalize=False,
+                                title="Confusion matrix, without normalization",
+                            )
+                            # Plot normalized confusion matrix
+                            self._plot_confusion_matrix(
+                                cm,
+                                self.class_names,
+                                normalize=True,
+                                title="Normalized confusion matrix",
+                            )
+                            plt.show()
+                else:
+                    if self.verbose:
+                        logger.info(
+                            "measures of error NOT evaluated for model:{}".format(
+                                self.kind_name
+                            )
                         )
-                        plt.show()
-        # todo TBD
-        elif self.type == "Regression":
-            pass
 
-        if self.measure:
-            if self.verbose:
-                logger.info(
-                    "measures of error evaluated for model: {}".format(self.learner)
-                )
-        else:
-            if self.verbose:
-                logger.info(
-                    "measures of error NOT evaluated for model:{}".format(self.learner)
-                )
+            # todo TBD
+            elif self.clf_type == "Regression":
+                pass
 
         return [y_pred, y_predp]
 
@@ -241,7 +247,7 @@ class Learner(pasoModel):
         logLoss = log_loss(y_act, y_pred_probability, eps=1e-15, normalize=True)
         if self.n_class == 2:
             # fix brier_loss
-#            brier_loss = brier_score_loss(y_act, y_pred_probability)
+            #            brier_loss = brier_score_loss(y_act, y_pred_probability)
             brier_loss = 0.0
             return {"brier_loss": brier_loss, "logloss": logLoss}
         else:
@@ -272,8 +278,12 @@ class Learner(pasoModel):
                 logger.info("Confusion matrix, without normalization")
 
         fig, ax = plt.subplots()
-        rcParams["figure.figsize"] = (self.x_size, self.y_size)
-        np.set_printoptions(precision=self.precision)
+
+        #        rcParams["figure.figsize"] = (self.x_size, self.y_size)
+        rcParams["figure.figsize"] = (6.0, 6.0)
+        #        np.set_printoptions(precision=self.precision)
+        np.set_printoptions(precision=3)
+
         im = ax.imshow(cm, interpolation="nearest", cmap=cmap)
         ax.figure.colorbar(im, ax=ax)
         # We want to show all ticks...